new code + updates

+++ compound intrest.py
@@ -1,6 +1,6 @@
-principle = int(input("Enter principle amount: "))
-rate = int(input("Enter interest rate: "))
-time = int(input("Enter time(years): "))
+principle = float(input("Enter principle amount: "))
+rate = float(input("Enter interest rate: "))
+time = float(input("Enter time(years): "))
 
 ci= principle * pow((1 + rate / 100), time) - principle
